LPS-55885 Moves code from EditLayoutAction to Layout Branches
