+++ portal-impl/src/com/liferay/portlet/stagingbar/action/EditLayoutBranchAction.java
@@ -19,14 +19,17 @@ import com.liferay.portal.NoSuchGroupException;
 import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.servlet.SessionErrors;
 import com.liferay.portal.kernel.servlet.SessionMessages;
+import com.liferay.portal.kernel.staging.StagingUtil;
 import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.security.auth.PrincipalException;
 import com.liferay.portal.service.LayoutBranchServiceUtil;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.service.ServiceContextFactory;
+import com.liferay.portal.theme.ThemeDisplay;
 import com.liferay.portal.util.PortalUtil;
 import com.liferay.portal.util.PortletKeys;
+import com.liferay.portal.util.WebKeys;
 import com.liferay.portlet.layoutsadmin.action.EditLayoutsAction;
 
 import java.util.HashMap;
@@ -37,6 +40,8 @@ import javax.portlet.ActionResponse;
 import javax.portlet.PortletConfig;
 import javax.portlet.RenderRequest;
 import javax.portlet.RenderResponse;
+
+import javax.servlet.http.HttpServletRequest;
 
 import org.apache.struts.action.ActionForm;
 import org.apache.struts.action.ActionForward;
@@ -71,6 +76,9 @@ public class EditLayoutBranchAction extends EditLayoutsAction {
 			else if (cmd.equals(Constants.DELETE)) {
 				deleteLayoutBranch(actionRequest);
 			}
+			else if (cmd.equals("select_layout_branch")) {
+				selectLayoutBranch(actionRequest);
+			}
 
 			if (SessionErrors.isEmpty(actionRequest)) {
 				SessionMessages.add(
@@ -84,10 +92,11 @@ public class EditLayoutBranchAction extends EditLayoutsAction {
 				data.put("preventNotification", Boolean.TRUE.toString());
 
 				SessionMessages.add(
-					actionRequest,
-					PortalUtil.getPortletId(actionRequest) +
-						SessionMessages.KEY_SUFFIX_REFRESH_PORTLET_DATA,
-					data);
+						actionRequest,
+						PortalUtil.getPortletId(actionRequest) +
+								SessionMessages.KEY_SUFFIX_REFRESH_PORTLET_DATA,
+						data
+				);
 			}
 
 			sendRedirect(actionRequest, actionResponse);
@@ -168,6 +177,23 @@ public class EditLayoutBranchAction extends EditLayoutsAction {
 				PortalUtil.getPortletId(actionRequest) +
 					SessionMessages.KEY_SUFFIX_PORTLET_NOT_AJAXABLE);
 		}
+	}
+
+	protected void selectLayoutBranch(ActionRequest actionRequest) {
+		HttpServletRequest request = PortalUtil.getHttpServletRequest(
+			actionRequest);
+
+		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
+			WebKeys.THEME_DISPLAY);
+
+		long layoutSetBranchId = ParamUtil.getLong(
+			actionRequest, "layoutSetBranchId");
+
+		long layoutBranchId = ParamUtil.getLong(
+			actionRequest, "layoutBranchId");
+
+		StagingUtil.setRecentLayoutBranchId(
+			request, layoutSetBranchId, themeDisplay.getPlid(), layoutBranchId);
 	}
 
 	protected void updateLayoutBranch(ActionRequest actionRequest)
