+++ portal-web/docroot/html/portlet/staging_bar/view_layout_branch_details.jsp
@@ -50,7 +50,7 @@ String stagingFriendlyURL = (String)request.getAttribute("view.jsp-stagingFriend
 						%>
 
 							<portlet:actionURL var="layoutBranchURL">
-								<portlet:param name="struts_action" value="/dockbar/edit_layouts" />
+								<portlet:param name="struts_action" value="/staging_bar/edit_layout_branch" />
 								<portlet:param name="<%= Constants.CMD %>" value="select_layout_branch" />
 								<portlet:param name="redirect" value="<%= stagingFriendlyURL %>" />
 								<portlet:param name="groupId" value="<%= String.valueOf(curLayoutBranch.getGroupId()) %>" />
