+++ portal-web/docroot/html/portlet/staging_bar/view_layout_set_branch_details.jsp
@@ -45,7 +45,7 @@ String stagingFriendlyURL = (String)request.getAttribute("view.jsp-stagingFriend
 						%>
 
 							<portlet:actionURL var="layoutSetBranchURL">
-								<portlet:param name="struts_action" value="/dockbar/edit_layouts" />
+								<portlet:param name="struts_action" value="/staging_bar/edit_layout_set_branch" />
 								<portlet:param name="<%= Constants.CMD %>" value="select_layout_set_branch" />
 								<portlet:param name="redirect" value="<%= stagingFriendlyURL %>" />
 								<portlet:param name="groupId" value="<%= String.valueOf(curLayoutSetBranch.getGroupId()) %>" />
